Move FormStep view into components dir
